+++ src/main/java/dev/saintho/mytly/api/v1/urls/dto/response/UrlPostResponse.java
@@ -1,4 +1,4 @@
-package dev.saintho.mytly.web.v1.dto.response;
+package dev.saintho.mytly.api.v1.urls.dto.response;
 
 import dev.saintho.mytly.domain.entity.Url;
 import lombok.Getter;
