+++ src/main/java/dev/saintho/mytly/service/UrlService.java
@@ -9,9 +9,9 @@ import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import dev.saintho.mytly.web.v1.dto.command.UrlDeleteByShortenedCommand;
-import dev.saintho.mytly.web.v1.dto.command.UrlShortCommand;
-import dev.saintho.mytly.web.v1.dto.query.UrlRedirectQuery;
+import dev.saintho.mytly.api.v1.urls.dto.command.UrlDeleteByShortenedCommand;
+import dev.saintho.mytly.api.v1.urls.dto.command.UrlShortCommand;
+import dev.saintho.mytly.api.v1.urls.dto.query.UrlRedirectQuery;
 import dev.saintho.mytly.domain.entity.Url;
 import dev.saintho.mytly.event.dto.UrlCreateEvent;
 import dev.saintho.mytly.exception.MytlyException;
